refactor(core): move macOS coordinate scale kernel into glass-core
Single source of truth for point<->pixel conversion, shared by glass-macos
(capture/window ops) and the forthcoming glass-a11y-macos reader so a11y
bounds and capture/input geometry cannot drift.
